Deleted TIM3 PWM output, only outputting PWM4 (The visible LED's).
Added multiple rates in in addition to an atomic subsystem that generates its own function.

Deleted functions that were just copy and pasted from the ST blocksets.

Added another tab with the option to connect STLink for flashing after compile. Added the appropriate hooks to the make_rtw_hook.

Removed ChibiOS as a target hardware device type. Rely on the built in supported devices.

Changed the compile make template to use the model name as the project name instead of "TestModel".